# PERF: Index.difference