- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.


asv continuous -f 1.1 upstream/main perf-index-difference -b index_object.SetDisjoint

```
       before           after         ratio
-      7.97±0.4ms       3.16±0.4ms     0.40  index_object.SetDisjoint.time_datetime_difference_disjoint
```

asv continuous -f 1.1 upstream/main perf-index-difference -b multiindex_object.Difference

```
       before           after         ratio
-      3.55±0.7ms       1.29±0.3ms     0.36  multiindex_object.Difference.time_difference('int')
-      3.79±0.5ms       1.20±0.2ms     0.32  multiindex_object.Difference.time_difference('datetime')
```