@rhshadrach 2023-10-20
This patch may have induced a performance regression. If it was a necessary behavior change, this may have been expected and everything is okay.

Please check the links below. If any ASVs are parameterized, the combinations of parameters that a regression has been detected for appear as subbullets.

 - [index_object.SetDisjoint.time_datetime_difference_disjoint](https://asv-runner.github.io/asv-collection/pandas/#index_object.SetDisjoint.time_datetime_difference_disjoint)
 - [multiindex_object.Difference.time_difference](https://asv-runner.github.io/asv-collection/pandas/#multiindex_object.Difference.time_difference)
   - [dtype='int'](https://asv-runner.github.io/asv-collection/pandas/#multiindex_object.Difference.time_difference?p-dtype=%27int%27)
   - [dtype='datetime'](https://asv-runner.github.io/asv-collection/pandas/#multiindex_object.Difference.time_difference?p-dtype=%27datetime%27)
   - [dtype='ea_int'](https://asv-runner.github.io/asv-collection/pandas/#multiindex_object.Difference.time_difference?p-dtype=%27ea_int%27)
   - [dtype='string'](https://asv-runner.github.io/asv-collection/pandas/#multiindex_object.Difference.time_difference?p-dtype=%27string%27)

Subsequent benchmarks may have skipped some commits. The link below lists the commits that are between the two benchmark runs where the regression was identified.

https://github.com/pandas-dev/pandas/compare/79067a76adc448d17210f2cf4a858b0eb853be4c...0bdbc44babac09225bdde02b642252ce054723e3